Move driver code into sub-directory
The new Makefile has been updated to allow for module compiling outside
of the Linux kernel directory tree. It uses a variable 'KDIR' to point
to the kernel source root directory. The variable has a default value
pointing the the local systems kernel src location but it can be
overridden using an exported environmental variable or as an argument
passes into the make command.