Reorder src/dst operations in xdma_perform_transaction
The logic being that if we have a long transaction for data that will
need to start writing to memory before its finished reading. But we want
to wait until we are sure that the transaction is complete. Then the
wait should be on the dst and that should be last.