DMA channel device-id used to filter xdma channel
The DMA channel node in the device tree has a "device-id" field which is
now used when filtering the DMA channels for use by the xdma driver.
This brings the drivers use of the device-id field in line with the
dmatest module in the linux-xlnx repo.

This may necessitate an update of the device tree to include the field
if not already included.

Example of the dma-channel node in the device tree:-

dma-channel@40400000 {
	compatible = "xlnx,axi-dma-mm2s-channel";
	interrupts = < 0 59 4 >;
	xlnx,datawidth = <0x20>;
	xlnx,device-id = <0x0>;
} ;