Update libxdma to use the new xdma header
This requires defining a dma_cookie_t type that is compatible to the real
<linux/dmaengine.h> type. The type is defined instead of including the
dmaengine header to free the user from having to install kernel headers on the
system, thus making the library easier to compile and install. The downside to
doing so is that dmaengine.h must always be monitored in case the definition of
dma_cookie_t is ever changed.